(1) Refactor, Create a new Stage1Configuration File with function (2) Refactor, move the stage 1 configuration from stage manager switch to the configuration file
